dom/indexedDB/test/test_third_party.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Test</title>

  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
    const BEHAVIOR_ACCEPT = 0;
    const BEHAVIOR_REJECTFOREIGN = 1;
    const BEHAVIOR_REJECT = 2;
    const BEHAVIOR_LIMITFOREIGN = 3;

    const testData = [
      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_ACCEPT, expectedResult: true },
      { host: "http://example.com", cookieBehavior: BEHAVIOR_ACCEPT, expectedResult: true },
      { host: "http://sub1.test2.example.org:8000", cookieBehavior: BEHAVIOR_ACCEPT, expectedResult: true },
      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_ACCEPT, expectedResult: true },

      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_REJECT, expectedResult: false },
      { host: "http://example.com", cookieBehavior: BEHAVIOR_REJECT, expectedResult: false },
      { host: "http://sub1.test2.example.org:8000", cookieBehavior: BEHAVIOR_REJECT, expectedResult: false },
      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_REJECT, expectedResult: false },

      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_REJECTFOREIGN, expectedResult: true },
      { host: "http://example.com", cookieBehavior: BEHAVIOR_REJECTFOREIGN, expectedResult: false },
      { host: "http://sub1.test2.example.org:8000", cookieBehavior: BEHAVIOR_REJECTFOREIGN, expectedResult: false },
      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_REJECTFOREIGN, expectedResult: true },

      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_LIMITFOREIGN, expectedResult: true },
      { host: "http://example.com", cookieBehavior: BEHAVIOR_LIMITFOREIGN, expectedResult: false },
      { host: "http://sub1.test2.example.org:8000", cookieBehavior: BEHAVIOR_LIMITFOREIGN, expectedResult: false },
      { host: "http://" + window.location.host, cookieBehavior: BEHAVIOR_LIMITFOREIGN, expectedResult: true },
    ];

    const iframe1Path =
      window.location.pathname.replace("test_third_party.html",
                                       "third_party_iframe1.html");
    const iframe2URL =
      "http://" + window.location.host +
      window.location.pathname.replace("test_third_party.html",
                                       "third_party_iframe2.html");

    let testIndex = 0;
    let openedWindow;

    // Cookie preference changes are only applied to top-level tabs/windows
    // when they are loaded. We need a window-proxy to continue the test.
    function openWindow() {
      SpecialPowers.pushPrefEnv({
        "set": [
          ["network.cookie.cookieBehavior", testData[testIndex].cookieBehavior],
        ],
      }, () => {
        openedWindow = window.open("third_party_window.html");
        openedWindow.onload = _ => {
          openedWindow.postMessage({
            source: "parent",
            href: iframe2URL,
            iframeUrl: testData[testIndex].host + iframe1Path,
          }, "*");
        };
      });
    }

    function messageListener(event) {
      // eslint-disable-next-line no-eval
      let message = eval(event.data);

      is(message.source, "iframe", "Good source");
      is(message.result, testData[testIndex].expectedResult, "Good result");

      openedWindow.close();

      if (testIndex < testData.length - 1) {
        testIndex++;
        openWindow();
        return;
      }

      SimpleTest.finish();
    }

    function runTest() {
      SimpleTest.waitForExplicitFinish();

      SpecialPowers.addPermission("indexedDB", true, document);

      window.addEventListener("message", messageListener);
      openWindow();
    }
  </script>

</head>

<body onload="runTest();">
</body>

</html>
